Reposition banner in tree
