+++ src/index.tsx
@@ -36,6 +36,7 @@ import getLibrary from 'utils/getLibrary'
 import AppziButton from 'components/AppziButton'
 import RadialGradientByChainUpdater from 'theme/RadialGradientByChainUpdater'
 import { nodeRemoveChildFix } from 'utils/node'
+import SideBanner, { BannerType } from 'components/SideBanner'
 
 // Node removeChild hackaround
 // based on: https://github.com/facebook/react/issues/11538#issuecomment-417504600
@@ -81,6 +82,7 @@ ReactDOM.render(
                 <Updaters />
                 <ThemeProvider>
                   <ThemedGlobalStyle />
+                  <SideBanner type={BannerType.ANNIVERSARY} />
                   <AppziButton />
                   <App />
                 </ThemeProvider>
